Improve the webhook error status code for not found errors

+++ components/org.wso2.carbon.identity.api.server.webhook.management/org.wso2.carbon.identity.api.server.webhook.management.v1/src/main/java/org/wso2/carbon/identity/api/server/webhook/management/v1/core/ServerWebhookManagementService.java
@@ -141,6 +141,11 @@ public class ServerWebhookManagementService {
     public void deleteWebhook(String webhookId) {
 
         try {
+            if (webhookManagementService.getWebhook(webhookId,
+                    CarbonContext.getThreadLocalCarbonContext().getTenantDomain()) == null) {
+                throw WebhookManagementAPIErrorBuilder.buildAPIError(Response.Status.NOT_FOUND,
+                        ERROR_NO_WEBHOOK_FOUND_ON_GIVEN_ID, webhookId);
+            }
             webhookManagementService.deleteWebhook(webhookId,
                     CarbonContext.getThreadLocalCarbonContext().getTenantDomain());
         } catch (WebhookMgtException e) {
@@ -157,6 +162,11 @@ public class ServerWebhookManagementService {
     public WebhookResponse activateWebhook(String webhookId) {
 
         try {
+            if (webhookManagementService.getWebhook(webhookId,
+                    CarbonContext.getThreadLocalCarbonContext().getTenantDomain()) == null) {
+                throw WebhookManagementAPIErrorBuilder.buildAPIError(Response.Status.NOT_FOUND,
+                        ERROR_NO_WEBHOOK_FOUND_ON_GIVEN_ID, webhookId);
+            }
             return getWebhookResponse(webhookManagementService.activateWebhook(webhookId,
                     CarbonContext.getThreadLocalCarbonContext().getTenantDomain()));
         } catch (WebhookMgtException e) {
@@ -173,6 +183,11 @@ public class ServerWebhookManagementService {
     public WebhookResponse deactivateWebhook(String webhookId) {
 
         try {
+            if (webhookManagementService.getWebhook(webhookId,
+                    CarbonContext.getThreadLocalCarbonContext().getTenantDomain()) == null) {
+                throw WebhookManagementAPIErrorBuilder.buildAPIError(Response.Status.NOT_FOUND,
+                        ERROR_NO_WEBHOOK_FOUND_ON_GIVEN_ID, webhookId);
+            }
             return getWebhookResponse(webhookManagementService.deactivateWebhook(webhookId,
                     CarbonContext.getThreadLocalCarbonContext().getTenantDomain()));
         } catch (WebhookMgtException e) {
@@ -189,6 +204,11 @@ public class ServerWebhookManagementService {
     public WebhookResponse retryWebhook(String webhookId) {
 
         try {
+            if (webhookManagementService.getWebhook(webhookId,
+                    CarbonContext.getThreadLocalCarbonContext().getTenantDomain()) == null) {
+                throw WebhookManagementAPIErrorBuilder.buildAPIError(Response.Status.NOT_FOUND,
+                        ERROR_NO_WEBHOOK_FOUND_ON_GIVEN_ID, webhookId);
+            }
             return getWebhookResponse(webhookManagementService.retryWebhook(webhookId,
                     CarbonContext.getThreadLocalCarbonContext().getTenantDomain()));
         } catch (WebhookMgtException e) {
